# openapi.yaml
openapi: 3.0.3
info:
  title: SplitBuddy API
  description: |
    Complete API specification for SplitBuddy expense splitting application.
    
    **Base URL**: http://localhost:420/api (local development)
    
    **Authentication**: Most endpoints require a Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: SplitBuddy API Support
servers:
  - url: http://localhost:420/api
    description: Local development server
  - url: http://localhost:4321/api
    description: Docker development server
...
paths:
...
  /friends/requests:
    post:
      tags:
        - Friends
      summary: Send friend request
      description: Send a friend request to another user
      operationId: sendFriendRequest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequestDto'
      responses:
        '201':
          description: Friend request sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /friends/requests/{requestId}:
    put:
      tags:
        - Friends
      summary: Respond to friend request
      description: Accept or reject a friend request
      operationId: respondToFriendRequest
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
          description: Friend request ID
        - name: response
          in: query
          required: true
          schema:
            type: string
            enum: [ACCEPTED, REJECTED]
          description: Response to the friend request
      responses:
        '200':
          description: Request updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /friends/{userId}:
    get:
      tags:
        - Friends
      summary: Get user's friends
      description: Get all friends for a specific user
      operationId: getFriends
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: List of friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendResponse'
  /friends/{userId}/pending-requests:
    get:
      tags:
        - Friends
      summary: Get pending friend requests
      description: Get all pending friend requests for a user
      operationId: getPendingRequests
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: List of pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendRequestResponse'
  /groups:
    post:
      tags:
        - Groups
      summary: Create group
      description: Create a new group with members
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '200':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups/me:
    get:
      tags:
        - Groups
      summary: Get my groups
      description: Get all groups for the current authenticated user
      operationId: getMyGroups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
  /groups/{id}:
    delete:
      tags:
        - Groups
      summary: Delete group
      description: Delete a group by ID
      operationId: deleteGroup
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Group ID
      responses:
        '204':
          description: Group deleted successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
...
    FriendRequestDto:
      type: object
      required:
        - senderId
        - receiverId
      properties:
        senderId:
          type: integer
          description: Sender's user ID
        receiverId:
          type: integer
          description: Receiver's user ID
        message:
          type: string
          description: Optional message
    FriendRequestResponse:
      type: object
      properties:
        id:
          type: integer
          description: Friend request ID
        senderName:
          type: string
          description: Sender's name
        senderEmail:
          type: string
          format: email
          description: Sender's email
        createdAt:
          type: string
          format: date-time
          description: Request creation timestamp
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, CANCELLED]
          description: Request status
        sender:
          $ref: '#/components/schemas/UserResponse'
        receiver:
          $ref: '#/components/schemas/UserResponse'
    FriendResponse:
      type: object
      properties:
        id:
          type: integer
          description: Friend's user ID
        name:
          type: string
          description: Friend's name
        email:
          type: string
          format: email
          description: Friend's email
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        name:
          type: string
          description: User's name
        email:
          type: string
          format: email
          description: User's email
    CreateGroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Group name
        memberIds:
          type: array
          items:
            type: string
          description: List of user IDs to add as members
    GroupResponse:
      type: object
      properties:
        id:
          type: integer
          description: Group ID
        name:
          type: string
          description: Group name
        createdBy:
          type: integer
          description: Creator's user ID
        createdByName:
          type: string
          description: Creator's name
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMemberResponse'
    GroupMemberResponse:
      type: object
      properties:
        id:
          type: integer
          description: Member's user ID
        name:
          type: string
          description: Member's name
        email:
          type: string
          format: email
          description: Member's email
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
        errors:
          type: object
          additionalProperties:
            type: string
          description: Field-specific validation errors